Avoid repeated typedefs
yyscan_t was already protected by YY_TYPEDEF_YY_SCANNER_T in
eval_l.c, so used this #define in eval_defs.h.  For ParseData,
added a new #define STAR_TYPEDEF_PARSEDATA so that it is
clear that this is a local patch.  This avoids compilation errors
with, for example, GCC 4.4.